criação do objeto tema e atualização do objeto postagem

+++ src/main/java/com/generation/blogpessoal/controller/PostagemController.java
@@ -1,6 +1,7 @@
 package com.generation.blogpessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,9 +16,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.generation.blogpessoal.model.Postagem;
 import com.generation.blogpessoal.repository.PostagemRepository;
+import com.generation.blogpessoal.repository.TemaRepository;
 
 import jakarta.validation.Valid;
 
@@ -29,6 +32,9 @@ public class PostagemController {
 	
 	@Autowired
 	private PostagemRepository postagemRepository;
+	
+	@Autowired
+	private TemaRepository temaRepository;
 	
 	
 	@GetMapping
@@ -76,7 +82,11 @@ public class PostagemController {
 	@ResponseStatus(HttpStatus.NO_CONTENT)
 	@DeleteMapping("/{id}")
 	public void delete(@PathVariable Long id) {
-		postagemRepository.deleteById(id);
+		  Optional<Postagem> postagem = postagemRepository.findById(id);
+		  if(postagem.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+	      postagemRepository.deleteById(id);
+	        
+	        /*No SQL: DELETE FROM tb_postagens WHERE id = id*/
 	}
 	
 }
